refactor(storage-resize-images): some refactoring
